msa로 수정된 파일

+++ src/main/java/kr/nyamnyam_kr/service/impl/RestaurantServiceImpl.java
@@ -30,9 +30,19 @@ public class RestaurantServiceImpl implements RestaurantService {
         return restaurantRepository.findById(id);
     }
 
+
     @Override
-    public void deleteById(Long id) {
-        restaurantRepository.deleteById(id);
+    public boolean deleteById(Long id) {
+        try {
+            if (restaurantRepository.existsById(id)) {
+                restaurantRepository.deleteById(id);
+                return true;
+            } else {
+                return false;
+            }
+        } catch (Exception e) {
+            return false;
+        }
     }
 
     @Override
